build-manifest: generate files in a separate directory and then merge it
This commit changes how build-manifest is invoked, stopping it from
generating the manifests directly into the downloads directory. Instead,
the files are generated in a separate directory and then merged back.

While there is not much point in doing so *right now*, this will be
crucial to implement smoke tests for the release, as we need to generate
a separate set of manifests with a different download URL to be able to
perform the smoke test. Having the generated output in a separate
directory will allow to easily discard those temporary manifests and
their signatures in a clean way.