smoke_test: ensure the release works with rustup before uploading
This commit changes the release process to execute a "smoke test" before
uploading the release to the destination buckets, by generating another
set of manifests pointing to a localhost server serving the files we're
about to upload.

After the localhost web server is up then rustup is invoked to download
the toolchain, and cargo is invoked to create a binary project and run
the hello world in it. If any of this fails the release is aborted.