config: add PROMOTE_RELEASE_BYPASS_STARTUP_CHECKS variable
This commit adds the PROMOTE_RELEASE_BYPASS_STARTUP_CHECKS enviornment
variable, which replaces the PROMOTE_RELEASE_ALLOW_MULTIPLE_TODAY
environment variable and adds more uses to it.

Namely, the new enviornment variable will allow to skip all the startup
checks that are meant to prevent an unwanted release from happening, not
just the one that prevents multiple releases in the same day.

The use for this will mainly be to re-run dev stable releases.