Switch dist artifact uploads to intelligent tiering
This should automatically put artifacts in cold storage as they decrease in use
over time. We do this for all artifacts, including documentation. Many of the
files uploaded for documentation aren't going to actually be eligible for colder
storage (there's a minimum size of 128KB) but in that case they'll just be
stored in standard storage as before.