Create and push rust-lang/rust tags for new stable releases
We could consider expanding this to beta's as well (or even nightly), but for
now this matches what the normal release process does well enough. It also won't
currently do anything in production (we need to configure the
PROMOTE_RELEASE_REPOSITORY with push credentials and PROMOTE_RELEASE_PUSH_TAGS),
but will still exercise the tag creation, which is good.

Code is intended to be somewhat extensible so that we can add Cargo tagging
pretty easily.