Automatic release promotion
This bumps branches on rust-lang/rust and creates the cargo branch at the
appropriate commit; the only missing component is the actual PR that will bump
the channel (src/ci/channel) to stable and cherry pick a fresh copy of the
release notes. I think it makes sense for now to leave that step to human hands:
it's pretty manual (need to track down the release notes, which might or might
not have merged already; need to check for last minute beta backports...). So
humans likely are involved anyway, at which point it's not really a huge win for
us to automate opening a PR or creating a temporary branch.

This automation already avoids the annoying bits and the bits that require
privileges (e.g., force pushing branches).