Recompress xz -> xz with tuned compression options
This adds support for recompressing the .xz tarballs in promote-release,
with a goal of moving rust-lang/rust CI back to producing "balanced"
profile binaries rather than the highly compressed ones currently
produced. This will cut rust-lang/rust CI times while keeping the
production artifacts in static.r-l.o equivalently compressed.

This PR doesn't enable the .xz recompression by default, that will be
done in a follow up simpleinfra push. Depending on the impact to
(especially) stable build times we may wish to bump the promote-release
production container from 8 cores to 72 cores (the next largest
container) so that we still finish in a reasonable amount of time.