Try to optimize runtime by sorting the recompression list
See comment in code for details. Currently our builds see 23/48 minutes
of the recompression runtime with less than full CPU utilization. The
hope is that this will help cut down on that time, so that we can more
efficiently make use of the available resources.

If this doesn't have any effect we'll consider the FIXME in the code as
well. xz compressed tarballs have metadata on the decompressed size
that's not currently readable by xz2 as well, in case file length is too
poor a proxy.