Invalidate release distribution on Fastly
We are working on serving Rust releases through Fastly as well as
CloudFront. The logic to invalidate cached released has therefore
been slightly reworked to handle both CDNs.

First, new environment variables have been added to configure the Fastly
integration. An API token is required as is the domain name that should
get invalidated.

Second, a temporary environment variable has been added that enables the
integration. This enables us to test the integration in the dev
environment without requiring any changes to the production environment.