Invalidate the cache on Fastly using surrogate keys
Fastly does not support wildcard invalidations through its URL-based
API, and also does not support invalidating a path prefix like `/dist/`.
We are therefore switching to surrogate keys. Requests that match the
path `^\/dist\/` get tagged with the surrogate key `/dist/*`, which is
then used when invalidating the cache after a release.

The tagging implementation can be found in rust-lang/simpleinfra#295.