Simplify the surrogate key for Fastly
In the original implementation, we simply used the path for
invalidations as the surrogate key. This worked fine when using the user
interface to invalidate the cache, but broke when trying to use Fastly's
API to do the same. While it is possible that we simply needed to
URL-encode the key, a safer implementation is to avoid the problem in
the first place by restricting the key to alphanumeric characters.