Restore override version and check for an update
Comparing the Rustup versions before and after the update does not work,
since the version in the GitHub repository is not bumped after a
release. We're going back to the override version so that we can force
rustup to self update, which we can check for without comparing version
numbers.

This reverts commit c9b96feda1929f03f5447d28e9c4aa7c3576943c.