Release nightly even if the commit is the same
It's convenient for people and tools if every nightly exists, as it
means people can pin any nightly without thinking. Sometimes,
due to CI issues, it can happen that no PR is merged in an entire day.
Even then, releasing nightly makes sense, as nightly is intended as a,
well, nightly snapshot of master, and if master doesn't change, we can
still snapshot it.