Properly select optimal XZ dictionary size
The recompressor will do GZip compression first, take note of the
decompressed size measured in the process, and use it to select an
optimal dictionary size for the XZ recompression, if it is to happen at
all. If XZ recompression is needed but GZip compression isn't, it will
perform a decompression solely to measure the decompressed size.